Modified starteam plugin to accept an additional parameter - destinationpath. This allows the user to target a specific destination folder in the jenkins workspace (which helps when checking out multiple repos). Also changed "scm" references to "instance" in the config.jelly page in order to work properly with the multiple scms plugin.
Unfortunately I was unable to check if the starteam unit/integration tests are still functioning as our local starteam repo uses encryption and the tests were throwing exceptions related to this. That said, I was able to test the changes in a running jenkins installation against a real repo and they seem to work properly.